+++ Server/Ebenezer/AISocket.cpp
@@ -248,7 +248,7 @@ void CAISocket::CloseProcess()
 void CAISocket::LoginProcess(char* pBuf)
 {
 	int index = 0;
-	float fReConnectEndTime = 0.0f;
+	double fReConnectEndTime = 0.0;
 	uint8_t zone = GetByte(pBuf, index);
 	// 0: first connect
 	// 1: reconnect
@@ -294,7 +294,7 @@ void CAISocket::LoginProcess(char* pBuf)
 				spdlog::info("AISocket::LoginProcess: sockets reconnected in under 2 minutes [sockets={}]",
 					_main->m_sReSocketCount);
 				_main->m_sReSocketCount = 0;
-				_main->m_fReConnectStart = 0.0f;
+				_main->m_fReConnectStart = 0.0;
 			}
 
 			if (_main->m_sReSocketCount == MAX_AI_SOCKET)
@@ -315,7 +315,7 @@ void CAISocket::LoginProcess(char* pBuf)
 				else
 				{
 					_main->m_sReSocketCount = 0;
-					_main->m_fReConnectStart = 0.0f;
+					_main->m_fReConnectStart = 0.0;
 				}
 			}
 		}
