+++ Server/Ebenezer/EbenezerDlg.h
@@ -247,21 +247,21 @@ public:
 	CKnightsSiegeWar		m_KnightsSiegeWar;
 
 	int16_t	m_sPartyIndex;
-	int16_t	m_sZoneCount;							// AI Server 재접속시 사용
-	int16_t	m_sSocketCount;							// AI Server 재접속시 사용
+	int16_t	m_sZoneCount;			// AI Server 재접속시 사용
+	int16_t	m_sSocketCount;			// AI Server 재접속시 사용
 	// sungyong 2002.05.23
 	int16_t	m_sSendSocket;
 	bool 	m_bFirstServerFlag;		// 서버가 처음시작한 후 게임서버가 붙은 경우에는 1, 붙지 않은 경우 0
 	bool 	m_bServerCheckFlag;
-	bool 	m_bPointCheckFlag;		// AI서버와 재접전에 NPC포인터 참조막기 (true:포인터 참조, false:포인터 참조 못함)
+	bool	m_bPointCheckFlag;		// AI서버와 재접전에 NPC포인터 참조막기 (true:포인터 참조, false:포인터 참조 못함)
 	int16_t	m_sReSocketCount;		// GameServer와 재접시 필요
-	float   m_fReConnectStart;	// 처음 소켓이 도착한 시간
-	int16_t	m_sErrorSocketCount;  // 이상소켓 감시용
+	double	m_fReConnectStart;		// 처음 소켓이 도착한 시간
+	int16_t	m_sErrorSocketCount;	// 이상소켓 감시용
 	// ~sungyong 2002.05.23
 
-	int m_iPacketCount;		// packet check
-	int m_iSendPacketCount;		// packet check
-	int m_iRecvPacketCount;		// packet check
+	int m_iPacketCount;				// packet check
+	int m_iSendPacketCount;			// packet check
+	int m_iRecvPacketCount;			// packet check
 
 	int m_nYear, m_nMonth, m_nDate, m_nHour, m_nMin, m_nWeather, m_nAmount;
 	int m_nCastleCapture;
@@ -289,13 +289,13 @@ public:
 	char	m_strBuyTitle[MAX_BBS_POST][MAX_BBS_TITLE];
 	char	m_strBuyMessage[MAX_BBS_POST][MAX_BBS_MESSAGE];
 	int		m_iBuyPrice[MAX_BBS_POST];
-	float	m_fBuyStartTime[MAX_BBS_POST];
+	double	m_fBuyStartTime[MAX_BBS_POST];
 
 	int16_t	m_sSellID[MAX_BBS_POST];
 	char	m_strSellTitle[MAX_BBS_POST][MAX_BBS_TITLE];
 	char	m_strSellMessage[MAX_BBS_POST][MAX_BBS_MESSAGE];
 	int		m_iSellPrice[MAX_BBS_POST];
-	float	m_fSellStartTime[MAX_BBS_POST];
+	double	m_fSellStartTime[MAX_BBS_POST];
 
 	// ~Yookozuna 2002.11.26 - 비러머글 남는 공지 --;
 	bool 	m_bPermanentChatMode;
